chore(app): merge app-dynamic-legacy with app
