--- packages/app-dynamic-legacy/.eslintignore
@@ -1,1 +1,0 @@
-public
