--- packages/app-dynamic-legacy/README.md
@@ -1,3 +1,0 @@
-# @janus-idp/app-dynamic-legacy
-
-This is an alternative to the `app-next` package. This package will use the current UI system rather than the upcoming declarative UI system.
