--- packages/app-dynamic-legacy/package.json
@@ -1,77 +1,0 @@
-{
-  "name": "app-dynamic-legacy",
-  "version": "0.0.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "bundled": true,
-  "publishConfig": {
-    "access": "public",
-    "main": "dist/index.esm.js",
-    "types": "dist/index.d.ts"
-  },
-  "backstage": {
-    "role": "frontend"
-  },
-  "scripts": {
-    "start": "janus-cli package start",
-    "build": "janus-cli package build",
-    "tsc": "tsc",
-    "clean": "backstage-cli package clean",
-    "test": "backstage-cli package test --passWithNoTests --coverage",
-    "lint": "backstage-cli package lint"
-  },
-  "dependencies": {
-    "@backstage/app-defaults": "^1.4.1",
-    "@backstage/catalog-model": "^1.4.1",
-    "@backstage/core-app-api": "^1.9.0",
-    "@backstage/core-components": "^0.13.3",
-    "@backstage/core-plugin-api": "^1.5.3",
-    "@backstage/integration-react": "^1.1.15",
-    "@backstage/plugin-api-docs": "^0.9.6",
-    "@backstage/plugin-catalog": "^1.12.0",
-    "@backstage/plugin-catalog-common": "^1.0.15",
-    "@backstage/plugin-catalog-graph": "^0.2.32",
-    "@backstage/plugin-catalog-import": "^0.10.1",
-    "@backstage/plugin-catalog-react": "^1.8.0",
-    "@backstage/plugin-github-actions": "^0.6.1",
-    "@backstage/plugin-org": "^0.6.10",
-    "@backstage/plugin-permission-react": "^0.4.14",
-    "@backstage/plugin-scaffolder": "^1.14.1",
-    "@backstage/plugin-search": "^1.3.3",
-    "@backstage/plugin-search-react": "^1.6.3",
-    "@backstage/plugin-tech-radar": "^0.6.6",
-    "@backstage/plugin-techdocs": "^1.6.5",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.0.15",
-    "@backstage/plugin-techdocs-react": "^1.1.8",
-    "@backstage/plugin-user-settings": "^0.7.5",
-    "@backstage/theme": "^0.4.1",
-    "@material-ui/core": "4.12.4",
-    "@material-ui/icons": "4.11.3",
-    "@openshift/dynamic-plugin-sdk": "^4.0.0",
-    "@scalprum/core": "0.6.0",
-    "@scalprum/react-core": "0.6.0",
-    "history": "5.3.0",
-    "lodash": "^4.17.21",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router": "6.16.0",
-    "react-router-dom": "6.16.0",
-    "react-use": "17.4.0"
-  },
-  "devDependencies": {
-    "@backstage/test-utils": "1.4.1",
-    "@janus-idp/cli": "*",
-    "@testing-library/jest-dom": "5.17.0",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/user-event": "14.5.1",
-    "@types/node": "18.16.16",
-    "@types/react-dom": "17.0.20",
-    "cross-env": "7.0.3"
-  },
-  "files": [
-    "dist"
-  ],
-  "configSchema": "config.d.ts"
-}
