--- packages/app-dynamic-legacy/src/App.test.tsx
@@ -1,5 +1,0 @@
-describe('App', () => {
-  it('should pass', async () => {
-    expect(true).toBe(true);
-  });
-});
