--- packages/app-dynamic-legacy/src/components/Root/Root.tsx
@@ -1,100 +1,0 @@
-import React, { PropsWithChildren } from 'react';
-
-import {
-  Link,
-  Sidebar,
-  sidebarConfig,
-  SidebarDivider,
-  SidebarGroup,
-  SidebarItem,
-  SidebarPage,
-  SidebarScrollWrapper,
-  SidebarSpace,
-  useSidebarOpenState,
-} from '@backstage/core-components';
-import { SidebarSearchModal } from '@backstage/plugin-search';
-
-import { makeStyles } from '@material-ui/core';
-import CreateComponentIcon from '@material-ui/icons/AddCircleOutline';
-import ExtensionIcon from '@material-ui/icons/Extension';
-import HomeIcon from '@material-ui/icons/Home';
-import LibraryBooks from '@material-ui/icons/LibraryBooks';
-import MenuIcon from '@material-ui/icons/Menu';
-import MapIcon from '@material-ui/icons/MyLocation';
-import SearchIcon from '@material-ui/icons/Search';
-import { ScalprumComponent } from '@scalprum/react-core';
-
-import LogoFull from './LogoFull';
-import LogoIcon from './LogoIcon';
-
-const useSidebarLogoStyles = makeStyles({
-  root: {
-    width: sidebarConfig.drawerWidthClosed,
-    height: 3 * sidebarConfig.logoHeight,
-    display: 'flex',
-    flexFlow: 'row nowrap',
-    alignItems: 'center',
-    marginBottom: -14,
-  },
-  link: {
-    width: sidebarConfig.drawerWidthClosed,
-    marginLeft: 24,
-  },
-});
-
-const SidebarLogo = () => {
-  const classes = useSidebarLogoStyles();
-  const { isOpen } = useSidebarOpenState();
-
-  return (
-    <div className={classes.root}>
-      <Link to="/" underline="none" className={classes.link} aria-label="Home">
-        {isOpen ? <LogoFull /> : <LogoIcon />}
-      </Link>
-    </div>
-  );
-};
-
-export const Root = ({ children }: PropsWithChildren<{}>) => (
-  <SidebarPage>
-    <Sidebar>
-      <SidebarLogo />
-      <SidebarGroup label="Search" icon={<SearchIcon />} to="/search">
-        <SidebarSearchModal />
-      </SidebarGroup>
-      <SidebarDivider />
-      <SidebarGroup label="Menu" icon={<MenuIcon />}>
-        {/* Global nav, not org-specific */}
-        <SidebarItem icon={HomeIcon} to="catalog" text="Home" />
-        <SidebarItem icon={ExtensionIcon} to="api-docs" text="APIs" />
-        <SidebarItem icon={LibraryBooks} to="docs" text="Docs" />
-        <SidebarItem icon={CreateComponentIcon} to="create" text="Create..." />
-        {/* End global nav */}
-        <SidebarDivider />
-        <SidebarScrollWrapper>
-          <SidebarItem icon={MapIcon} to="tech-radar" text="Tech Radar" />
-        </SidebarScrollWrapper>
-      </SidebarGroup>
-      <SidebarSpace />
-      <SidebarDivider />
-      <SidebarGroup
-        label="Settings"
-        icon={
-          <ScalprumComponent
-            scope="janus.dynamic-frontend-plugin"
-            module="UserSettings"
-            importName="UserSettingsSignInAvatar"
-          />
-        }
-        to="/settings"
-      >
-        <ScalprumComponent
-          scope="janus.dynamic-frontend-plugin"
-          module="UserSettings"
-          importName="Settings"
-        />
-      </SidebarGroup>
-    </Sidebar>
-    {children}
-  </SidebarPage>
-);
