--- packages/app-dynamic-legacy/src/components/Root/index.ts
@@ -1,1 +1,0 @@
-export { Root } from './Root';
