--- packages/app-dynamic-legacy/src/index.tsx
@@ -1,1 +1,0 @@
-import('./bootstrap');
