--- packages/app-dynamic-legacy/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@backstage/cli/config/tsconfig.json",
-  "include": ["src"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "outDir": "../../dist-types/packages/app-dynamic-legacy",
-    "rootDir": "."
-  }
-}
