--- packages/app-dynamic-legacy/turbo.json
@@ -1,14 +1,0 @@
-{
-  "extends": ["//"],
-  "pipeline": {
-    "start": {
-      "cache": false,
-      "persistent": true,
-      "dependsOn": ["^build"]
-    },
-    "tsc": {
-      "outputs": ["../../dist-types/packages/app-dynamic-legacy/**"],
-      "dependsOn": ["^tsc"]
-    }
-  }
-}
