+++ packages/app/package.json
@@ -1,14 +1,22 @@
 {
   "name": "app",
   "version": "0.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
   "private": true,
   "bundled": true,
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.esm.js",
+    "types": "dist/index.d.ts"
+  },
   "backstage": {
     "role": "frontend"
   },
   "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
+    "start": "janus-cli package start",
+    "build": "janus-cli package build",
     "tsc": "tsc",
     "clean": "backstage-cli package clean",
     "test": "backstage-cli package test --passWithNoTests --coverage",
@@ -17,7 +25,6 @@
   "dependencies": {
     "@backstage/app-defaults": "^1.4.4",
     "@backstage/catalog-model": "^1.4.3",
-    "@backstage/cli": "^0.23.0",
     "@backstage/core-app-api": "^1.11.0",
     "@backstage/core-components": "^0.13.6",
     "@backstage/core-plugin-api": "^1.7.0",
@@ -42,15 +49,21 @@
     "@backstage/theme": "^0.4.3",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
+    "@openshift/dynamic-plugin-sdk": "^4.0.0",
+    "@scalprum/core": "^0.6.0",
+    "@scalprum/react-core": "^0.6.0",
     "history": "^5.3.0",
+    "lodash": "^4.17.21",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router": "^6.11.2",
+    "react-router": "^6.17.0",
     "react-router-dom": "^6.17.0",
     "react-use": "^17.4.0"
   },
   "devDependencies": {
+    "@backstage/cli": "0.23.0",
     "@backstage/test-utils": "1.4.4",
+    "@janus-idp/cli": "1.2.0",
     "@testing-library/jest-dom": "5.17.0",
     "@testing-library/react": "12.1.5",
     "@testing-library/user-event": "14.5.1",
@@ -72,5 +85,6 @@
   },
   "files": [
     "dist"
-  ]
+  ],
+  "configSchema": "config.d.ts"
 }
