+++ packages/app/src/App.tsx
@@ -1,111 +1,115 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { Navigate, Route } from 'react-router';
 
-import { createApp } from '@backstage/app-defaults';
 import { FlatRoutes } from '@backstage/core-app-api';
 import { AlertDisplay, OAuthRequestDialog } from '@backstage/core-components';
-import { apiDocsPlugin, ApiExplorerPage } from '@backstage/plugin-api-docs';
-import {
-  CatalogEntityPage,
-  CatalogIndexPage,
-  catalogPlugin,
-} from '@backstage/plugin-catalog';
+import { ApiExplorerPage } from '@backstage/plugin-api-docs';
+import { CatalogEntityPage, CatalogIndexPage } from '@backstage/plugin-catalog';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
-import {
-  CatalogImportPage,
-  catalogImportPlugin,
-} from '@backstage/plugin-catalog-import';
-import { orgPlugin } from '@backstage/plugin-org';
+import { CatalogImportPage } from '@backstage/plugin-catalog-import';
 import { RequirePermission } from '@backstage/plugin-permission-react';
-import { ScaffolderPage, scaffolderPlugin } from '@backstage/plugin-scaffolder';
 import { SearchPage } from '@backstage/plugin-search';
-import { TechRadarPage } from '@backstage/plugin-tech-radar';
 import {
   TechDocsIndexPage,
-  techdocsPlugin,
   TechDocsReaderPage,
 } from '@backstage/plugin-techdocs';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
-import { UserSettingsPage } from '@backstage/plugin-user-settings';
+
+import { ScalprumProvider } from '@scalprum/react-core';
 
 import { apis } from './apis';
 import { entityPage } from './components/catalog/EntityPage';
 import { Root } from './components/Root';
 import { searchPage } from './components/search/SearchPage';
+import DynamicRoot from './DynamicRoot';
+import DynamicRootContext from './DynamicRoot/DynamicRootContext';
 
-const app = createApp({
-  apis,
-  bindRoutes({ bind }) {
-    bind(catalogPlugin.externalRoutes, {
-      createComponent: scaffolderPlugin.routes.root,
-      viewTechDoc: techdocsPlugin.routes.docRoot,
-    });
-    bind(apiDocsPlugin.externalRoutes, {
-      registerApi: catalogImportPlugin.routes.importPage,
-    });
-    bind(scaffolderPlugin.externalRoutes, {
-      registerComponent: catalogImportPlugin.routes.importPage,
-    });
-    bind(orgPlugin.externalRoutes, {
-      catalogIndex: catalogPlugin.routes.catalogIndex,
-    });
-  },
-});
+const App = () => {
+  const { AppProvider, AppRouter, dynamicRoutes } =
+    useContext(DynamicRootContext);
+  return (
+    <AppProvider>
+      <AlertDisplay />
+      <OAuthRequestDialog />
+      <AppRouter>
+        <Root>
+          <FlatRoutes>
+            <Route path="/" element={<Navigate to="catalog" />} />
+            <Route path="/catalog" element={<CatalogIndexPage />} />
+            <Route
+              path="/catalog/:namespace/:kind/:name"
+              element={<CatalogEntityPage />}
+            >
+              {entityPage}
+            </Route>
+            <Route path="/docs" element={<TechDocsIndexPage />} />
+            <Route
+              path="/docs/:namespace/:kind/:name/*"
+              element={<TechDocsReaderPage />}
+            >
+              <TechDocsAddons>
+                <ReportIssue />
+              </TechDocsAddons>
+            </Route>
+            <Route path="/api-docs" element={<ApiExplorerPage />} />
+            <Route
+              path="/catalog-import"
+              element={
+                <RequirePermission permission={catalogEntityCreatePermission}>
+                  <CatalogImportPage />
+                </RequirePermission>
+              }
+            />
+            <Route path="/search" element={<SearchPage />}>
+              {searchPage}
+            </Route>
+            <Route path="/catalog-graph" element={<CatalogGraphPage />} />
+            {dynamicRoutes.map(({ Component, path, ...props }) => (
+              <Route
+                key={path}
+                path={path}
+                element={<Component {...props} />}
+              />
+            ))}
+          </FlatRoutes>
+        </Root>
+      </AppRouter>
+    </AppProvider>
+  );
+};
 
-const AppProvider = app.getProvider();
-const AppRouter = app.getRouter();
+const manifestLocation = 'http://localhost:8004/plugin-manifest.json';
+const remotePluginId = 'janus.dynamic-frontend-plugin';
 
-const routes = (
-  <FlatRoutes>
-    <Route path="/" element={<Navigate to="catalog" />} />
-    <Route path="/catalog" element={<CatalogIndexPage />} />
-    <Route
-      path="/catalog/:namespace/:kind/:name"
-      element={<CatalogEntityPage />}
+const AppRoot = () => {
+  return (
+    <ScalprumProvider
+      pluginSDKOptions={{
+        pluginLoaderOptions: {
+          postProcessManifest: manifest => {
+            return {
+              ...manifest,
+              loadScripts: manifest.loadScripts.map(
+                script => `http://localhost:8004/${script}`,
+              ),
+            };
+          },
+        },
+      }}
+      config={{
+        [remotePluginId]: {
+          manifestLocation,
+          name: remotePluginId,
+        },
+      }}
     >
-      {entityPage}
-    </Route>
-    <Route path="/docs" element={<TechDocsIndexPage />} />
-    <Route
-      path="/docs/:namespace/:kind/:name/*"
-      element={<TechDocsReaderPage />}
-    >
-      <TechDocsAddons>
-        <ReportIssue />
-      </TechDocsAddons>
-    </Route>
-    <Route path="/create" element={<ScaffolderPage />} />
-    <Route path="/api-docs" element={<ApiExplorerPage />} />
-    <Route
-      path="/tech-radar"
-      element={<TechRadarPage width={1500} height={800} />}
-    />
-    <Route
-      path="/catalog-import"
-      element={
-        <RequirePermission permission={catalogEntityCreatePermission}>
-          <CatalogImportPage />
-        </RequirePermission>
-      }
-    />
-    <Route path="/search" element={<SearchPage />}>
-      {searchPage}
-    </Route>
-    <Route path="/settings" element={<UserSettingsPage />} />
-    <Route path="/catalog-graph" element={<CatalogGraphPage />} />
-  </FlatRoutes>
-);
+      <DynamicRoot apis={apis}>
+        <App />
+      </DynamicRoot>
+    </ScalprumProvider>
+  );
+};
 
-const App = () => (
-  <AppProvider>
-    <AlertDisplay />
-    <OAuthRequestDialog />
-    <AppRouter>
-      <Root>{routes}</Root>
-    </AppRouter>
-  </AppProvider>
-);
-
-export default App;
+export default AppRoot;
