+++ packages/app/src/DynamicRoot/dynamic-modules-registry.json
