+++ packages/app/src/DynamicRoot/index.ts
