+++ packages/app/src/index.tsx
@@ -1,5 +1,3 @@
-import '@backstage/cli/asset-types';
-
 import React from 'react';
 import ReactDOM from 'react-dom';
